Beginning to create Edit function.

+++ src/components/pages/Projects.jsx
@@ -1,9 +1,11 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable no-unused-vars */
-import { supabase } from "../../createCliente";
+import { supabase } from "../../services/createCliente";
 import { useEffect, useState } from "react";
 import Cardproject from "../project/Cardproject";
 import { Flex, Wrap, WrapItem} from "@chakra-ui/react";
+import EditarProjeto from "../project/EditarProjeto";
+import { Link } from "react-router-dom";
 
 
 
@@ -25,12 +27,26 @@ const Projects = () => {
 
   // Delete by ID
   async function deleteById(id){
-   const { data , error} = await supabase.from("projeto").delete().eq('id', id)
-    if(error){
-     console.log(error.message);
+    if(confirm("Deseja apagar este projeto?")){
+      const { data , error} = await supabase.from("projeto").delete().eq('id', id)
+      if(error){
+       console.log(error.message);
+      }
+      getProjects();
+    } 
+    return;
     }
-    getProjects();
-  }
+    // Update by ID
+    async function updateById (id){
+      const {data, error} = await supabase.from ("projeto").select("*").eq("id", id)
+      if(error){
+        console.log(error.message)
+      }
+       setProjects(data)
+       
+
+    } 
+   
 
     
 
@@ -39,7 +55,8 @@ const Projects = () => {
 
         <Flex justifyContent="space-around" wrap="wrap" gap={3}> 
          {projects.map((project) =>(
-          <Cardproject key={project.id} id={project.id} nome={project.nome} orcamento={project.orcamento} categoria={project.id_tipo} deleteById={deleteById} />
+          <Cardproject key={project.id} id={project.id} nome={project.nome} orcamento={project.orcamento} categoria={project.id_tipo}
+           deleteById={deleteById} updateById={updateById} data={projects}/>
          ))}
        </Flex> 
         </div>
